+++ src/main/kotlin/com/justai/jaicf/profit/scenario/MainScenario.kt
@@ -8,6 +8,8 @@ import com.justai.jaicf.model.scenario.Scenario
 import com.justai.jaicf.profit.ProfitCalculator
 import com.justai.jaicf.profit.model.Product
 import kotlinx.serialization.json.content
+import kotlinx.serialization.json.double
+import kotlinx.serialization.json.float
 import kotlinx.serialization.json.int
 
 object MainScenario: Scenario() {
@@ -55,9 +57,9 @@ object MainScenario: Scenario() {
                     val firstUnit = slots["first_unit"]
                     val secondUnit = slots["second_unit"] ?: firstUnit
 
-                    context.session["first"] = Product(firstAmount?.value?.int ?: 1, firstPrice!!.value.int, firstUnit!!.value.content)
+                    context.session["first"] = Product(firstAmount?.value?.double ?: 1.0, firstPrice!!.value.int, firstUnit!!.value.content)
                     context.session["second"] = secondPrice?.let {
-                        Product(secondAmount?.value?.int ?: 1, secondPrice.value.int, secondUnit!!.value.content)
+                        Product(secondAmount?.value?.double ?: 1.0, secondPrice.value.int, secondUnit!!.value.content)
                     }
 
                     reactions.go("calculate")
@@ -116,7 +118,7 @@ object MainScenario: Scenario() {
 
                         val first = context.session["first"] as Product
                         context.session["second"] = Product(
-                            secondAmount?.value?.int ?: 1,
+                            secondAmount?.value?.double ?: 1.0,
                             secondPrice!!.value.int,
                             secondUnit?.value?.content ?: first.unit
                         )
